CRUD Functions

+++ demo-project/app/Http/Controllers/API/CourseApiController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Http\JsonResponse;
 Use App\Models\Course;
 
@@ -16,7 +17,7 @@ class CourseApiController extends Controller
     {
         $validatedData = $request->validate([
             'course' => 'required|string|max:255',
-            'description' => 'required|string|max',
+            'description' => 'required|string',
         ]);
 
         $course = new Course();
@@ -30,25 +31,37 @@ class CourseApiController extends Controller
     public function show($id)
     {
         return Course::find($id);
+        if(!`$course`){
+            return response()->json(['message'=>'Course not found'],Response::HTTP_NOT_FOUND);
+        }
         return response()->json($course);
     }
 
-    //Updating data
+    //Updating data 
     public function update(Request $request, $id)
     {
         $course = Course::find($id);
-        $course->course = $request->course;
-        $course->description = $request->description;
+        if(!$course){
+            return response()->json(['message'=>'Course not found'],Response::HTTP_NOT_FOUND);
+        }
+        $validatedData = $request->validate([
+            'course' => 'string|max:255',
+            'description' => 'string',
+        ]);
+        $course->fill($request->all());
         $course->save();
-        return response()->json($course);
+        return response()->json(['message'=>'Course has been updated succesfully','course'=>$course],Response::HTTP_OK);
     }
 
     //Deleting data
     public function destroy($id)
     {
         $course = Course::find($id);
+        if(!$course){
+            return response()->json(['message'=>'Course not found'],Response::HTTP_NOT_FOUND);
+        }
         $course->delete();
-        return response()->json(null);
+        return response()->json(['message'=>'Course has been deleted succesfully'],Response::HTTP_OK);
     }
 
 
@@ -56,5 +69,6 @@ class CourseApiController extends Controller
     public function index()
     {
         return Course::all();
+        return response()->json($courses);
     }
 }
